services/tripgraph-service/pyproject.toml
[project]
name = "breeze-tripgraph-service"
version = "0.1.0"
description = "Breeze TripGraph Routing Engine — CSA + A* multi-modal routing for India"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "gunicorn>=21.2.0",
    "asyncpg>=0.29.0",
    "redis[hiredis]>=5.0.0",
    "scipy>=1.12.0",
    "numpy>=1.26.0",
    "apscheduler>=3.10.0",
    "aiokafka>=0.10.0",
    "pydantic>=2.6.0",
    "pydantic-settings>=2.1.0",
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
    "opentelemetry-exporter-otlp>=1.22.0",
    "httpx>=0.27.0",
    "alembic>=1.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-benchmark>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
